Resolve circular dependency hell
Unit tests were failing because of a circular dependency between
session <-> settings.
