+++ ui/src/session.ts
@@ -8,6 +8,8 @@ import * as notification from "./notification";
 import * as remote from "./remote";
 import { Appearance, CoCo, Settings } from "./settings";
 
+import { createValidationStore, ValidationStatus } from "./validation";
+
 // TYPES
 
 export interface Session {
@@ -125,3 +127,65 @@ export const updateCoCo = (coco: CoCo): void =>
     // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
     settings: { ...get(settings), coco },
   });
+
+const VALID_SEED_MATCH = /^[\w\d]{54}@[\w\d.]*:[\d]{1,5}$/;
+
+const checkSeedUniqueness = (seed: string): Promise<boolean> => {
+  return Promise.resolve(!get(settings).coco.seeds.includes(seed));
+};
+
+export const seedValidation = createValidationStore(
+  {
+    format: {
+      pattern: VALID_SEED_MATCH,
+      message: "This is not a valid seed address",
+    },
+  },
+  [
+    {
+      promise: checkSeedUniqueness,
+      validationMessage: "This seed already exists",
+    },
+  ]
+);
+
+let restartNotificationShown = false;
+
+const showRestartNotification = (): void => {
+  // The restart notification is sticky and will stay visible until the user
+  // clicks the "Restart" button, no need to show it multiple times.
+  if (restartNotificationShown) return;
+
+  notification.error(
+    "Restart the app to connect to your new seed",
+    false,
+    "Restart",
+    () => {
+      // TODO(rudolfs): implement proxy/app restart
+      console.log("restart");
+    },
+    true
+  );
+
+  restartNotificationShown = true;
+};
+
+export const addSeed = async (seed: string): Promise<boolean> => {
+  // This has to be awaited contrary to what tslint suggests, because we're
+  // running async remote validations in in the background. If we remove the
+  // async then the seed input form will have to be submitted twice to take any
+  // effect.
+  await seedValidation.validate(seed);
+  if (get(seedValidation).status !== ValidationStatus.Success) return false;
+
+  updateCoCo({ seeds: [...get(settings).coco.seeds, seed] });
+  showRestartNotification();
+  return true;
+};
+
+export const removeSeed = (seed: string): void => {
+  updateCoCo({
+    seeds: get(settings).coco.seeds.filter((x: string) => x !== seed),
+  });
+  showRestartNotification();
+};
